Fix endNodes having edges in included strings
To make an example: consider "test" and "testing". "test" is included
in "testing", so if we create "test" nodes first and then go for "testing"
, while finding it's shared prefix , we might run into a node with a parent edge
of "EWO" and an evaluation of "test", which will cause problems because it'll
be considered the main node of the rest of the word (which is "ing") in the function
`createWord()`. so if we run into a "EWO" node which can evaluate
the prefix ("test" for this example) , we simply return the parent node
instead of itself.